часть 1ого, toString не выводит имена классов
