+++ Homework2/DesktopComputer.java
@@ -1,0 +1,17 @@
+package Homework2;
+
+abstract public class DesktopComputer extends ComputerImpl { //Шаблон настольного компьютера
+
+    public DesktopComputer(String maker, String model, CPU cpu, RAM ram, HDD hdd) {
+        super(maker, model, cpu, ram, hdd);
+    }
+
+    public String toString() {
+        StringBuilder sb = new StringBuilder();
+        sb.append("(Desktop Computer):\r\n");
+        for (Component component : components) {
+            sb.append(component.printComponent() + "\r\n");
+        }
+        return sb.toString();
+    }
+}
