+++ Homework2/PortableComputer.java
@@ -1,0 +1,17 @@
+package Homework2;
+
+abstract public class PortableComputer extends ComputerImpl { //Шаблон портативного компьютера
+
+    public PortableComputer(String maker, String model, CPU cpu, RAM ram, HDD hdd) {
+        super(maker, model, cpu, ram, hdd);
+    }
+
+    public String toString() {
+        StringBuilder sb = new StringBuilder();
+        sb.append("(Portable Computer):\r\n");
+        for (Component component : components) {
+            sb.append(component.printComponent() + "\r\n");
+        }
+        return sb.toString();
+    }
+}
